+++ test/scenarios/driver/openstack/molecule/default/tests/test_default.py
@@ -1,0 +1,28 @@
+import os
+
+import testinfra.utils.ansible_runner
+
+testinfra_hosts = testinfra.utils.ansible_runner.AnsibleRunner(
+    os.environ['MOLECULE_INVENTORY_FILE']).get_hosts('all')
+
+
+def test_hostname(SystemInfo):
+    assert 'instance-1-default' == SystemInfo.hostname
+
+
+def test_etc_molecule_directory(File):
+    f = File('/etc/molecule')
+
+    assert f.is_directory
+    assert f.user == 'root'
+    assert f.group == 'root'
+    assert f.mode == 0o755
+
+
+def test_etc_molecule_ansible_hostname_file(File):
+    f = File('/etc/molecule/instance-1-default')
+
+    assert f.is_file
+    assert f.user == 'root'
+    assert f.group == 'root'
+    assert f.mode == 0o644