+++ test/scenarios/driver/openstack/molecule/multi-node/tests/test_baz.py
@@ -1,0 +1,28 @@
+import os
+
+import testinfra.utils.ansible_runner
+
+testinfra_hosts = testinfra.utils.ansible_runner.AnsibleRunner(
+    os.environ['MOLECULE_INVENTORY_FILE']).get_hosts('baz')
+
+
+def test_hostname(SystemInfo):
+    assert 'instance-2-multi-node' == SystemInfo.hostname
+
+
+def test_etc_molecule_directory(File):
+    f = File('/etc/molecule')
+
+    assert f.is_directory
+    assert f.user == 'root'
+    assert f.group == 'root'
+    assert f.mode == 0o755
+
+
+def test_etc_molecule_ansible_hostname_file(File):
+    f = File('/etc/molecule/instance-2-multi-node')
+
+    assert f.is_file
+    assert f.user == 'root'
+    assert f.group == 'root'
+    assert f.mode == 0o644