+++ test/unit/test_config.py
@@ -31,6 +31,7 @@ from molecule.dependency import gilt
 from molecule.driver import dockr
 from molecule.driver import lxc
 from molecule.driver import lxd
+from molecule.driver import openstack
 from molecule.driver import static
 from molecule.driver import vagrant
 from molecule.lint import ansible_lint
@@ -124,6 +125,19 @@ def test_driver_property_is_lxd(molecule_driver_lxd_section_data,
     assert isinstance(config_instance.driver, lxd.Lxd)
 
 
+@pytest.fixture
+def molecule_driver_openstack_section_data():
+    return {'driver': {'name': 'openstack'}, }
+
+
+def test_driver_property_is_openstack(molecule_driver_openstack_section_data,
+                                      config_instance):
+    config_instance.merge_dicts(config_instance.config,
+                                molecule_driver_openstack_section_data)
+
+    assert isinstance(config_instance.driver, openstack.Openstack)
+
+
 def test_driver_property_is_static(molecule_driver_static_section_data,
                                    config_instance):
     config_instance.merge_dicts(config_instance.config,
@@ -164,7 +178,7 @@ def test_driver_property_raises(molecule_driver_invalid_section_data,
 
 
 def test_drivers_property(config_instance):
-    x = ['docker', 'lxc', 'lxd', 'static', 'vagrant']
+    x = ['docker', 'lxc', 'lxd', 'openstack', 'static', 'vagrant']
 
     assert x == config_instance.drivers
 
@@ -354,7 +368,7 @@ def test_molecule_file():
 
 
 def test_molecule_drivers():
-    x = ['docker', 'lxc', 'lxd', 'static', 'vagrant']
+    x = ['docker', 'lxc', 'lxd', 'openstack', 'static', 'vagrant']
 
     assert x == config.molecule_drivers()